requirements.txt
Pillow==8.3.2
pkg_resources==0.0.0
pytesseract==0.3.8
selenium==3.141.0
urllib3==1.26.6
